medical_buddy_agent/requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
xmltodict==0.13.0
wikipedia==1.4.0